Create JF_parent_paper_recreation.R
Added R script for attempting to implement the parent paper's code and results. Consolidated the parent paper's code from three files in its GitHub repo: "MEPS.R", "twostageSL.R", and "Estimators.R". Made the adjustment to remove a (seemingly) unnecessary variable from the test set ("X"), removed saving the MAE/MSE/Rsq results to a file, added functionality for the code to run across multiple local cores in parallel, and added file paths for the train and test data to this repo. The rest of the code is unchanged to the best of my knowledge.