Update JF_parent_paper_recreation_revised.R
Reverted two-stage SL algorithms back to their original form in the parent paper's code, as the changes I made were not able to run on my computer. My results reflect the code of the parent paper (although the algorithm sets in the parent paper's code do not exactly match what they published), except for changing the cross-fold validation parameter to 2 instead of 10 for computational feasibility.